[engine] Support arbitrary specs as input in legacy:list
This change is a quick cleanup to add test coverage for the legacy list command (which parses build files and then outputs all valid addresses found).

The plan is still to throw this entrypoint away, but for now I'm using it as a base to experiment with the Parser implementation for existing build files.

- Add CmdLineSpecParser-ing to list.
- Normalize Single/Descendant/Siblings to all return lists of product outputs.
- Add coverage tests for list.

Testing Done:
https://travis-ci.org/pantsbuild/pants/builds/112311362

Bugs closed: 2980

Reviewed at https://rbcommons.com/s/twitter/r/3511/