Extract a BuildGraph interface
Extract a `BuildGraph` interface and rename the existing impl to `MutableBuildGraph` in order to allow for an immutable implementation that will raise 'NotImplemented' for mutable methods.

This in service of [#2979](https://github.com/pantsbuild/pants/issues/2979), which hopes to fill this interface with an alternate implementation. That code currently lives in `engine.exp`, but as part of the first alternate implemention, we hope to promote a portion of it out of `engine.exp` shortly after 1.0.0.

Testing Done:
https://travis-ci.org/pantsbuild/pants/builds/112742003

Bugs closed: 2985

Reviewed at https://rbcommons.com/s/twitter/r/3514/