Make synthetic jar_library targets dependencies of android_binary.
There are a few other tiny fixes here to make all the
android tests go green - I am going to attempt to bring
the android backend into contrib.

Previously the jars were dependents of the library they belonged
to. But that was before the isolated_compile - with that
strategy the jars were not making it onto the classpath of the
android_binary. In a correctness sense, this is a better
representation of the graph.

Unfortunately, this commit does not quite cover all the dependency
inferral that the android backend had before isolated compile.

AndroidLibrary targets are just ways to attach resources to jars
and optionally define include/exclude patterns. So any
AndroidBinary thats depends on an AndroidLibrary should
get the libraries classes.jar on its classpath for free.
But as you can see with the change to hello_with_library,
more than one AndroidLibrary in a dependency chain
and you need to add the 3rdparty jars to the binary
target as well.

A nice followup would be to refactor unpack_libraries
and dx_compile to associate the unpacked_libraries
products with AndroidBinary targets instead of
AndroidLibrary. An issue will be opened.

With this commit, all the android tests go green and
./pants bundle examples/src/android::` runs succeessfully.