Move the android backend into contrib.
This creates the new contrib module. A followup
is going to remove any other special-casing the
android backend collected over the years.

This has to ship with the followup CR that
removes the android special casing from
src, test and build-support. But I broke
it up to be easier on reviewers.

This is mostly a pure move. There are a couple points of
interest.

* I created a new xml_test_base python_library.
  It was previously embedded within a python_test. But
  that class is imported in the contrib module. It was working
  before by some transitive magic.