Remove the API public from ImportJarsMixin
Doesn't look like this is an error - so i am
removing the public. I would like to convert
android_library to use the new paylaod_field.JarsField
but I don't know if I will have the time. For now
it looks okay for the contrib module to use a
non-public API.