Clone jars rather than mutating them during ivy resolve
The excludes (for 'provided'/round-trip deps) added to jars during ivy resolve are only useful in that scope, but because they remain in the graph, they can end up affecting the excludes written during `./pants publish` (and probably also the runtime performance of classpath calculation).

- Made JarDependency immutable by extending `datatype`
- Switched mutators to the `jar.copy(**replacements)` API
- Fixup the build dictionary to support `namedtuple`/`datatype` subclasses, which don't have `__init__`

While this is technically a breaking change, I think it's easy to make an argument for closing a loophole that allows for strange sideeffects in the build graph.

Testing Done:
https://travis-ci.org/pantsbuild/pants/builds/113190835

Bugs closed: 2655

Reviewed at https://rbcommons.com/s/twitter/r/3203/