Add classes.jar from unpacked libraries as a dependency of Android binaries.
Android targets generally need all the components from their closure
in the tool steps. With isolated compile. the classes.jar was on the
classpath of the AndroidLibrary but was not on the classpath of
the binary itself. This commit injects the classes.jar as a
JarLibrary dependency on all AndroidBinaries that have the
associated AndroidLibrary in its closure.

This now allows all the example apps to build and run, and preserves
all the ordering guarantees from the global strategy.

This is not themost graceful way to do it, but I am trying to get
the Android backend into contrib and all tests green. A goal would
be to protect that state with adding the SDK to the CI.