Add the Android SDK dependencies.
This tooka multi-hour investigation since
the workunits swallowed the useful error and
instead sent me on a wild-goose chase of
Subprocess invocations.

But anyway, the libraries are needed in order to
run the aapt tool - a fundamental building
block of android builds. Travis has a general
android config but that is used
to run pure android builds- not something
like Pants.