Add the Android SDK to the linux CI and turn on Android tests.
This installs and caches the Android SDK on linux CI
runs. The debug.keystore used to be installed along
with the SDK - but I had to recreate it for the CI.
That directory is cached so it should only run once
or if a layer changes.

The Android tests will now be run in the contrib shard.