Reimplement list options to support appendg.
To append a single value to a list-valued option, just "set" the option
to that value, as with action='append'.

./pants --foo=1 --foo=2

To append several values, add a plus sign in front of the value:

./pants --foo=+[1,2]

To replace the existing value, just set to a list:

./pants --foo=[1,2]

Note that this syntax works with config and env values as well.  This
means that you can append to the value specified in, say, pants.ini,
but also override it entirely, if you need to.

This endows list_option with the only useful feature of action='append',
so that will soon be deprecated.  In fact, action='append' is now implemented
as a list_option under the covers, so replacing all uses of it should be
straightforward.

In order to support lists of (scalar) values other than strings, this
introduces a member_type registration arg, which is only valid when
type=list_option.

This change was more complex than you'd think, as it required refactoring
Parser._compute_value. This method is now clearer and better documented
than before.

Some behavior has changed slightly. E.g., previously we validated
values even if they weren't used (e.g., the config, when it was overridden
by a cmd-line flag). Now we only validate the actual final value. But
these changes are either positive or neutral, in terms of their impact,
and they make the code more straightforward.