allow list-owners to accept multiple source files and output JSON
Testing Done:
Tested locally and via Travis CI: https://travis-ci.org/pantsbuild/pants/builds/113938458

Based on the comments in https://rbcommons.com/s/twitter/r/2755/ it looks like the original intention for this method was to accept multiple sources and return JSON

I left the output for one source file in the original format but I'm not sure if this is necessary, we could just change it to always output JSON?

I also left files that don't exist out of the JSON output.  We could possibly change this so an empty set of values is returned in the dictionary instead?

Bugs closed: 3009

Reviewed at https://rbcommons.com/s/twitter/r/3534/