[engine] Implement support for recursive path globs
This fills in the blanks left behind in the PathGlobs implementation. It now supports all forms of `filespec` as implemented by `wrapped_globs`. PathGlobs are projected from Sources objects in the new engine to implement the capture of Paths and FilesContent for `Sources` objects.

* Propagate the `errno` when enhancing exceptions in ProjectTree.
* Add fs test to cover PathGlobs.create.
* Rename PathGlob to PathLiteral.
* Make DirectoryListing "optional" in the sense that a non-existent entry is treated as empty.
* Implement support for single and double (recursive) wildcards in PathGlobs.

Attached is a rendering of the recursive glob for:

    ./pants run src/python/pants/engine/exp/examples:viz-fs -- \
      . \
      walk \
      'examples/tests/scala/org/pantsbuild/**/*.scala'

Testing Done:
https://travis-ci.org/pantsbuild/pants/builds/114646055

Bugs closed: 3015

Reviewed at https://rbcommons.com/s/twitter/r/3540/