Record the compile classpath used to compile jvm targets.
This change makes `jvm_compile` write text files storing the
classpath for each jvm target it compiles to:

    .pants.d/compile/zinc/<sha>/<target.id>/<sha>/classes/compile_classpath/<target.id>.txt

Since these files are written to the classes/ directory, they end
up being bundled into jvm binaries. This can be useful to determine
how an artifact was built after the fact.

This behavior is enabled by default, and can be disabled with:

    --no-compile-zinc-capture-classpath