Implementing scoped dependencies and classpath intransitivity.
This adds two fields to all targets: `scope` and `_transitive`.

These control how downstream tasks construct the set of targets
used for generating the classpath (in the case of JVM-related
tasks). This can be used to control similar behavior for other
(non-JVM) tasks, but this patch does not attempt to implement
all possible uses.

It is up to the individual task implementations to read and respect
these scopes. This can be facilitated by a couple utility methods,
such as BuildGraph.closure() and Target.closure(), which now take
in `include_scopes`, `exclude_scopes`, and `respect_intransitive`
parameters. The default values for all three of these parameters
results in the previously existing behavior of closure().

This patch modifies jvm tasks so that jvm compiles, runs, binaries,
and bundles respect scopes.

This adds a small number scopes to start with; more can easily be
added. See the Scopes class, where these are defined as constants.

  * DEFAULT - used all the time
  * COMPILE - used only at compile time
  * RUNTIME - used only at run-time (test, run, binary)
  * TEST - used only at test-time

It is possible to use these scopes and the transitive flag to
emulate the "provided" scope found in other build systems (such as
Maven, Gradle, and IntelliJ) like so:

  ```
  jar_library(name='my-provided-target',
    jars=[...],
    # Only available at test-time and compile-time.
    scope='compile test',
  )

  java_library(name='lib',
    dependencies=[
      intransitive(':my-provided-target'),
    ],
  )
  ```

For convenience, when applying a scope to a target is done rarely
or inconsistently, it may be less verbose to use the `scoped()`
alias like so:

  ```
  jar_library(name='my-provided-target',
    jars=[...],
  )

  java_library(name='lib',
    dependencies=[
      scoped(':my-provided-target',
             scope='compile test', transitive=False),
    ],
  )
  ```

This alias simply constructs an intermediary target with the
appropriate scope and transitive arguments, and returns its address
for inclusion in a dependencies list.