Only bootstrap the zinc worker pool if there is work to do
Currently, jvm compile will always create a worker pool. This updates JvmCompile so that it only bootstraps the pool if there are invalid targets. It also renames compile_chunk to do_compile as compile_chunk implies chunking behavior that is no longer used.

Testing Done:
Wrote an integration test that checks for the presence and abscence of the workunit label in output and made it pass. CI away on the PR.

Bugs closed: 3034

Reviewed at https://rbcommons.com/s/twitter/r/3559/