Add the Android SDK to the linux CI and turn on Android tests.
This installs and caches the Android SDK on linux CI
runs. The debug.keystore used to be installed along
with the SDK - but I had to recreate it for the CI.
That directory is cached so it should only run once
or if a layer changes.

The Android tests will now be run in the contrib shard.

In 3531 I maintained the '-a' flag as a way to skip
running Android tests, even now that it is moved to
contrib. I did that because since the OSX CI is not
container based it would have to redownload the entire
SDK each run. So Android tests (and the install-android-sdk
script) are only being run on the Linux CI.

Testing Done:
This is dependent on CR 3530 and 3531 landing - which move android into contrib.

A CI run that installs the SDK into the container and runs the Android test suite is green:
https://travis-ci.org/pantsbuild/pants/pull_requests

Bugs closed: 937, 2993, 3006, 3012

Reviewed at https://rbcommons.com/s/twitter/r/3538/