Fix unicode parsing of ini files.
We encountered an annoying problem upgrading pants in our repo.
Apparently we had a unicode left-quotation mark (\u2018) inside a
line comment. The line comment happened to be in a list. This
caused the option parsing to fail when trying to call
`value.startswith('[')`, which generated a cryptic error about not
being able to decode the string with ascii. This was especially
troublesome because there was no indication of which option the
error was in (and trying to get python to print out the string
caused it to fail with the same error just trying to format the
string for inclusion in a print message).