Fix several problems with golang thrift generation
1. Services mistakenly required some special thing called
_service_deps, which didn't exist.  This was copypasta from one
of the other thrift generators.

2. Go packages are always a single level.  The physical source files
can be located with "import paths", which are the directories leading
up to the go files which comprise a package. We want the file
foo/bar.thrift to generate foo/ttypes.go, with "package bar" at the
top.  You could just generate "ttypes.go", but then you have a problem
when multiple thrift files use the same package.  Apache's thrift
generator is supposed to handle this with the package_prefix option,
but from reading the source code, it appears that this is mostly not
hooked up (it changes the import paths in the {svc}-remote.go file for
services, but that's it).  Maybe I don't understand what it's supposed
to do?  Anyway, we need to manually relocate the generated dir from
the root to the appropriate subdirectory.

3. And after all that, Apache's thrift generator generates some these
{svc}-remote.go files for each service, all in the "main" package.
You can't have files from two different packages in the same directory
(subdirs are fine), but Pants's golang stuff (_symlink_lib) seems to
assume that you want to do that.  Since we don't really care about
these (afaict, they're basically just cute command-line clients for
the service), we'll just consider them as not sources, and not symlink
them.