[engine] Calculate legacy target sources using the engine
This review replaces `sources` fields on targets reified in the ExpGraph with a helper object that creates a FilesetWithSpec that uses the engine to compute the Paths and FileContent. This means that once construction of the Graph switches over, all BUILD files and source hashes will be cached.

This is in support of [#3058](https://github.com/pantsbuild/pants/issues/3058).

* Replace glob implementations with capturing implementations and create FilesetWithSpecs.
* Add `filemap` legacy command to demonstrate that we are able to compute paths.
* Mutate a single ParseContext in LegacyParser rather than recreating for each path.
* Normalize SourcesField to always consume a FilesetWithSpec object.
* Move the implementation of file-content fetching to FilesetWithSpec, to allow our alternate implementation to use the engine to get FileContent.
* Finished removing `num_chunking_units` (after [r/3474](https://rbcommons.com/s/twitter/r/3474/)), which was forcing sources in cases where we otherwise didn't need to.

Testing Done:
https://travis-ci.org/pantsbuild/pants/builds/118376108

Bugs closed: 3058, 3076

Reviewed at https://rbcommons.com/s/twitter/r/3592/