Prepare 0.0.79
While release the 0.0.78 release, I got a 500 error during an upload, and aborted not knowing whether it had succeeded. Retrying the upload failed, because the file existed, and even after deleting the file would not allow for a reupload at the same version.

So: bump to 0.0.79 in order to retry.

Testing Done:
https://travis-ci.org/pantsbuild/pants/builds/118727771
https://travis-ci.org/pantsbuild/pants/builds/118730468

Bugs closed: 3100

Reviewed at https://rbcommons.com/s/twitter/r/3619/