Fix some bad three-state logic in thrift_linter.
It was relying on a default of None in a boolean option
to detect if the option was explicitly provided or not.
But we never want boolean options to be None, as three
state logic is pointlessly complicated and a pending
change relies on this).

This change replaces that logic with a proper check.