Fix, document, or mark xfail tests that fail in Jenkins.
The junit tests fail when Java 8 is available,
but not the default.  These were likely not running at all in
Travis, but Jenkins has both JDK7 and 8 available, so it exposed
the issues.

The distribution integration test was failing because the distribution
home path was a symlink.  `realpath`ing the directory in the test util
code fixes the issue.

The export integration test makes assumptions about subinvocations
of pants that are not correct.  In particular, it assumes that the
default ivy cache path will be used.  This breaks when
PANTS_IVY_CACHE_DIR is set.  I left a comment indicating as much.

Testing Done:
CI is green: https://travis-ci.org/pantsbuild/pants/builds/120139430

Bugs closed: 3123

Reviewed at https://rbcommons.com/s/twitter/r/3632/