Fix a bug in OptionTracker.
Previously, if an option set in config (for example) had no hardcoded
default then the option tracker wouldn't see it as overridden.

This commit modifies the integration test to expose this problem,
and fixes it.

That said, I assume the previous logic was there for a reason, so
if this change is naive or otherwise wrong - please point it out!