Import zincutils
Re-import zincutils in order to make it easier to change. The motivation for moving this out was that this code might be ported to native code, which would be easier accomplished externally. That turned out to be unnecessary, and there are upcoming changes to the zinc/sbt version used in the pants repo that are more easily accomplished and tested with our full test suite in place.

This was roughly:
- `cp zincutils/zincutils src/python/pants/backend/jvm/zinc`
- `cp zincutils/tests to tests/python/pants_test/backend/jvm/zinc`
- Fixup imports, apply style fixes.

Testing Done:
https://travis-ci.org/pantsbuild/pants/builds/121269857

Bugs closed: 3156

Reviewed at https://rbcommons.com/s/twitter/r/3657/