Rename the global lock from pants.run to pants.workdir.file_lock
This change renames  the global lock in pants to keep it from conflicting with implementations prior to https://rbcommons.com/s/twitter/r/3633/

In prior implementations (using the python lockfile module) the presence of the file itself signals the lock.
After the change to use OwnerPrintingInterProcessFileLock which uses the fastener's library, the file will be left behind after the lock
is released.  This causes problems if you later invoke an older version of pants in a workspace after using the updated version of pants
(for example, which switching back to an older branch in the repo.)

Testing Done:
CI is running at https://travis-ci.org/pantsbuild/pants/builds/121803515

Bugs closed: 3163

Reviewed at https://rbcommons.com/s/twitter/r/3668/