Change some option defaults.
These are based on common pants.ini configs at Square, Foursquare and Twitter.

The nature of these options is such that changing the default is unlikely to have
adverse impact on anyone not currently setting them explicitly in pants.ini,
and we'd like to make them more useful before we freeze them in 1.0.