Resolve symlinks when generating sdists.
Previously pants would link symlinks as-is into sdist chroots.  Combined
with the fact pants depends on an old setuptools that does not follow
symlinks when creating sdists, this led to sdists with dangling links.
Control the behavior by having pants copy into the sdist chroot instead
of linking to ensure a valid final sdist with reolved files.