Require explicitly requesting cache reads during integration tests.
Some tests implicitly relied on the fact that
their integration tests wouldn't read the build cache.
But that truth depends on the state of the invoked
pants.ini.

My repro on current master is:

PANTS_CONFIG_OVERRIDE=pants.cache.ini ./pants test contrib/scrooge::

The tests fail because it reads the cache, as declared
in pants.ini.cache. The pants_run invocation
declares a --config-files flag, but the option
system (and the integration test library copies in the env)
concatenates it with any files declared by
PANTS_CONFIG_OVERRIDE. So in the above repro,
the pants.cache.ini turns caching on for all tasks.
This in turn has the scrooge tests read the cache
and not run the goals, e.g. causing the test runs
to pass when they are expected to fail.

In general, I think that using a temp workdir for an integration
test implies that the test was written assuming a clean
environment for its execution. I experimented with turning
cache reads off for all tests, but the ci runtime was
untenably increased. Plus, there was no good way to
opt-in to using the cache in that formulation.

Any test that relies on reading the cache now has to
opt-in by passing cache_read=True when calling a
pants_run method.