Some fixes to make tests more robust around jvm_options.
This came out of a discarded change to actually set the jvm_options
to be non-empty. This exposed various issues in tests that assume
that the default jvm_options are empty.

This commit keeps the fixes for those issues, so that if in the future
we do change the jvm_options defaults, the tests will continue to work.