Some fixes to make tests more robust around jvm_options.
This came out of a discarded change that actually set the jvm_options
to be non-empty, which exposed various issues in tests that assume
that the default jvm_options are empty.

This commit salvages the fixes for those issues, so that if in the future
we do change the jvm_options defaults, the tests will continue to work.

Note: this change does not modify the current jvm_options default, which
remains empty. It merely enables us to do so in the future.

Testing Done:
CI passes: https://travis-ci.org/pantsbuild/pants/builds/123574259

Reviewed at https://rbcommons.com/s/twitter/r/3706/