Include ivy options in ivy fingerprints
Options like --soft-excludes are not included in the cache key used by the ivy task mixin. Since the cache key is reused as the key for the resolve, this means for instance, that the bundle task might reuse the result of a resolve with --soft-excludes even though it expected excludes to be applied. It also means that associated subsystem options such as the jar dependency management default target are ignored. And, that the task implementation version is ignored.

This converts IvyResolveFingerprintStrategy to a TaskIdentityFingerprintStrategy subclass rather than a FingerprintStrategy subclass. I think it's possible to reduce this to just the ivy mixin related elements, but that started getting a little complicated, so I backed it out.

Testing Done:
CI away on PR. Ran a bundle / resolve pair that previously reused the same cached resolve now correctly have separate resolves.

Bugs closed: 3234

Reviewed at https://rbcommons.com/s/twitter/r/3729/