Fixup `PythonChrootTest.test_thrift_issues_2005`.
Previously, the test walked a pex chroot to do a tally of files.  This
approach was flawed since a pex chroot object rooted at a (temporary)
path can have files under the root that are not owned by the chroot
object. In order for the chroot object to own a file, it has to be added
via its API.  Its just these files that make it into a frozen pex.  In
the case of pants pex chroots, the chroot dir is used as a convenient
temp-dir base behind pex's back (files added to the temp-dirs are not
added to the pex chroot object).  This change fixes the chroot
inspection code to act against a clean copy of the chroot object's view
of the chroot.