Use NailgunTask's Java distribution consistently.
A few NailgunTasks need the JDK's tools.jar, and were fetching
its path directly via DistributionLocator, instead of using
the distribution it will later execute with.

In practice, today, these all end up being the same distribution.
But in the future, if a particular task restricts its choice of
distribution then we could end up running in one JDK with the
tools.jar of another JDK on our classpath.