Results of running Python buildgen over Pants codebase.
Foursquare is opensourcing our Buildgen implementation.
The buildgen module was refactored to be per-language,
and I imported it into a contrib module and ran it over pants.

We hope to release the source early this week - but in the meantime
a good test after the OSS-ing refactor was to run it over the
pants codebase.

The invocation was:

  ./pants buildgen src:: tests:: testsprojects:: examples:: \
          pants-plugins:: 3rdparty:: contrib::

Four concessions did I make to get the incantation to run:

* The old contrib/buildgen was not included - it is already inlined
  in the buildgen plugin itself
* Did not include the contrib/go backend.
    * (except to fix up the twitter.common move - see below)
    * The buildgen implementations were not playing well together.
* Did not include the 'test_antlr_builder' and 'test_thrift_namespace_packages'
   * These can be made to work with buildgen by creating
     buildgen_target_bags - bags that aggregate targets for things
     like codegen.
   * But since they are both marked 'xxx' I decided to see if people
     thought that was worth the time or maybe those tests should be
     deleted.
* Move twitter.commons from 3rdparty/python/twitter/common into plain
  3rdparty/python
    * It was just easier this way - could probably be worked around
      if someone feels strongly.

At least one consequence for merging this patch:
    * Buildgen will remove targets from bags if the targets
      have their own definition in that BUILD file.
    * I cleaned up the now-empty bags where I saw them. See
      contrib/android/src/python/pants/contrib/android/tasks
      for what I mean.

This was mechanically done for the most part - and anything
that we change from review will just get changed back if we ever
run buildgen again (which I hope is the case). So this is kind of
an all-or-nothing patch, with the exception of the twitter.common
move.