Remove overlap with 1.0.0-rc3.
The changelog is accurate in terms of releases but is not
accurate in respect to master. The specific problem commit
is 3ae0f77513f2da45b0348913d0417d829aaad86e in master.

It was cherry-picked into the 1.0.0-rc3 and 1.0.0 releases.
But commits that landed after it have never been included
in a release.

The changelog helper correctly identified it as within the range
between the last commit and HEAD - the commit had been
released but with a different SHA since it was cherry-picked
into the new branch.

IMO, the 1.0.0 and RCs should have maintained their own CHANGELOGs
and the last release on master should have been 0.0.82.
But others disagree and that ship has sailed.

I simply removed the duplicate RB commit from the list and
updated the CHANGELOG. This should not be a problem again
until we make a 1.1.0 branch and perhaps the policies will be
fleshed out by then.

I published the docs locally and inspected the output, will
include a screen shot.