[engine] Eagerly compute/cache sources in the engine
In order to cache globs execution and file hashes, this review moves to computing an `EagerFilesetWithSpecs` class in the engine (pre-fork, from the perspective of the daemon).

While this will have additional overhead in the case of a cold build, it allows more to be cached, it avoids invoking the engine once per target, and, importantly: it's simpler.

- Make `FilesetWithSpecs` abstract, and subclass with eager and lazy variants (which improves pickleability/indirection).
- Rename `LegacyBuildGraphNode` to `LegacyTarget` (still confusing, but no longer overloaded with `exp.nodes.Node`).
- Eagerly compute a FilesetWithSpecs for each LegacyTarget, which will cause them to be cached.
- Rename the `legacy.globs` module to `legacy.structs` and move TargetAdaptor there.
- Wrapped `bundles` and `sources` parameters via a generic `Field` abstraction.
- Outlawed nested excludes, and fixed the filemap tests to account for https://rbcommons.com/s/twitter/r/3840/

----

As tasks begin to migrate to consuming the engine API directly (for example: `./pants list` is a request for `an Address for each member of a DescendantAddresses object`), more work can be done pre-fork.

Testing Done:
https://travis-ci.org/pantsbuild/pants/builds/129266296

Bugs closed: 3328

Reviewed at https://rbcommons.com/s/twitter/r/3818/