Making the permissions of temporary files configurable.
Unintuitively, the file permissions temporary files are created
impact the file permissions of non-temporary files across many
parts of the build process. This is due to the common pattern of
creating temporary files and directories, then moving them over
their ultimate destinations atomically.

We addresses this issue partially in:

  https://rbcommons.com/s/twitter/r/3420/

However, there appear to be other places that this effect crops up.

This patch adds the ability to set default file permissions used
by contextutil when creating temporary files and directories.

Since pretty much everything (including subsystems) depend directly
or indirectly on utils, I was unfortunately not able to just put
these options directly into a subsystem. Instead I made a simple
class that contains configuration values that behaves similarly
to a subsystem, and an associated Task that populates it with data
from the actual options system.