Implement options scope name deprecation.
Allows an Optionable to specify that its current scope is
a new name for an older, deprecated scope.

Takes option values from both scopes, but issues a warning if
any values were explicitly specified on the old scope.

Note: Removes an update() method on OptionValueContainer that was
never used.  Adds instead an augment() method with similar semantics,
but that takes an OptionValueContainer instead of an attr dict.