Make pants a good example of Go contrib usage.
This fixes the pants repo to be `buildgen` compatible and `buildgen`
tested. To do so, extraneous and `buildgen` illegal Go source roots used
for tests are eliminated and `buildgen` is added to CI to ensure
compliance with its strictures. Along the way a misconception about how
`go_remote_library` works is corrected in a test: in short, there is no
choice about where a `go_remote_library` BUILD file lives, it must live
at the remote repo "root".